day13

+++ day13/src/oop/inherit4/File.java
@@ -2,14 +2,14 @@ package oop.inherit4;
 
 public class File {
 	private String filename;
-	private int filesize;
+	private float filesize;
 	public String getFilename() {
 		return filename;
 	}
 	public void setFilename(String name) {
 		this.filename = filename;
 	}
-	public int getFilesize() {
+	public float getFilesize() {
 		return filesize;
 	}
 	public void setFilesize(int size) {
