+++ day13/src/oop/inherit4/Hwp.java
@@ -1,8 +1,8 @@
 package oop.inherit4;
 
 public class Hwp extends File {
-	private int pagesize;
-	public int getPagesize() {
+	private float pagesize;
+	public float getPagesize() {
 		return pagesize;
 	}
 	public void setPagesize(int pagesize) {
